# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "World UZ",
        "description": "Documentation for World Uz Company",
        "contact": {
            "email": "rustamovramziddin7@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/field": {
            "get": {
                "tags": [
                    "Field"
                ],
                "summary": "Return only fields ",
                "description": "Get all data",
                "operationId": "9c6e7ee64d1737d64e61a9b2115f9dd1",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Field"
                ],
                "summary": "Post a new data",
                "description": "Post new University  data",
                "operationId": "b388943d11eb5ffe9e390c1959230bd8",
                "requestBody": {
                    "description": "Pass Country   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "university_id",
                                    "country_id",
                                    "name",
                                    "category",
                                    "price",
                                    "duration",
                                    "description"
                                ],
                                "properties": {
                                    "university_id": {
                                        "type": "number",
                                        "format": "int",
                                        "example": "1"
                                    },
                                    "country_id": {
                                        "type": "number",
                                        "format": "int",
                                        "example": "2"
                                    },
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Computer science"
                                    },
                                    "category": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Master"
                                    },
                                    "price": {
                                        "type": "text",
                                        "format": "string",
                                        "example": "$2000"
                                    },
                                    "duration": {
                                        "type": "text",
                                        "format": "string",
                                        "example": "12 monthy"
                                    },
                                    "description": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "something"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/field/{field}": {
            "get": {
                "tags": [
                    "Field"
                ],
                "summary": "Get one ",
                "description": "Return all date related to ID{field id}",
                "operationId": "dd3dd9aa6f363b474f205aec28f1ddb0",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Field"
                ],
                "summary": "Get one and Delete related to field id",
                "description": "Return  date related to ID of the field",
                "operationId": "d167794c520092daa1ee47f1d5ea8048",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "put field id and try to delete ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/v1/fields/{field}/field": {
            "get": {
                "tags": [
                    "Field"
                ],
                "summary": "Get one with all the relation data ",
                "description": "Return all date related to ID{field id}",
                "operationId": "2af48e09e37a71f75dc3f9af85b7945d",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "ID of the field ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}